Write relation info index files (WIP).
Currently completely wrong; endDocument() shouldn't try to write the
final files, this just means we're done with this one term in this doc.
Instead we should write a tmp file like we do with FI and remember the
offsets in this file for each doc. Then at the end we can convert this
file to per-doc relations info files.